// src/dorc-web/src/apis/dorc-api/models/DeploymentRequestAttemptApiModel.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type {
    DeploymentResultAttemptApiModel,
} from './';

/**
 * @export
 * @interface DeploymentRequestAttemptApiModel
 */
export interface DeploymentRequestAttemptApiModel {
    /**
     * @type {number}
     * @memberof DeploymentRequestAttemptApiModel
     */
    Id?: number;
    /**
     * @type {number}
     * @memberof DeploymentRequestAttemptApiModel
     */
    DeploymentRequestId?: number;
    /**
     * @type {number}
     * @memberof DeploymentRequestAttemptApiModel
     */
    AttemptNumber?: number;
    /**
     * @type {string}
     * @memberof DeploymentRequestAttemptApiModel
     */
    StartedTime?: string | null;
    /**
     * @type {string}
     * @memberof DeploymentRequestAttemptApiModel
     */
    CompletedTime?: string | null;
    /**
     * @type {string}
     * @memberof DeploymentRequestAttemptApiModel
     */
    Status?: string | null;
    /**
     * @type {string}
     * @memberof DeploymentRequestAttemptApiModel
     */
    Log?: string | null;
    /**
     * @type {string}
     * @memberof DeploymentRequestAttemptApiModel
     */
    UserName?: string | null;
    /**
     * @type {Array<DeploymentResultAttemptApiModel>}
     * @memberof DeploymentRequestAttemptApiModel
     */
    ComponentResults?: Array<DeploymentResultAttemptApiModel> | null;
}